Initial LED Module.
This is a project specific module and has yet to be tested on hardware. It takes a single write instruction via API function that requires a row and a mode number. The row determines which set of 5 leds to paint (upper or lower) and the mode value's bits are read and serve as high low signals for each element in the led array.