Initial I2C Module commit.
This is the first, untested pass at an i2c module. It is relatively generic at the moment but will need a lot of work. The main goal now is to get it working for Terpsichore's 8-channel DAC. Need to decide on whether to run this on a timer and to feed a buffer from another thread, or to simply have this queue a write command as needed. Since we'll be dealing with smapled external CV (and internal actually), it'll need to be very responsive. Still needs functional testing.