Refactored Led / Leds modules.
Separated out the actual engine of the led module that interacts with the GPIOs of the MCU. It's important I need to stipulate that this is a direct MCU to LED implementation. In the near future it might be better served that we just refactor this as a GPIO output module for flipping states. Leds module is now terpsichore specific that does the logic.