GPIO Module added.
This module needs to be tested, but it's designed to configure GPIO I/Os based on the DT inputs under the Zephyr,User tag. There is currently one callback for all inputs, which we'll want to figure out, but we can deal with that post-testing. This module should be able to handle all I/Os across different modules. For absolutely time-critical changes, we can still defer to the Zephyr API and gpio_pin_set directly.