ADC Module added.
Generic module to handle reading of ADC channels. Needs to be tested, but this module should run on a timer set by the user and broadcasts readings. Will have to figure out the corners as we test, and need to make sure we're efficient enough here. For Terpsichore in particular, we'll need to handle both potentiometers and CV inputs, both of which have different requirements, both in terms of resolution and sampling rates.